adds download projects to dashboard
